Lifetimes and tests
There has been some very slight modifications in how
`implement_dispatch` macro handles its arguments, which make it now able
to handle lifetimes as type arguments. This still comes with a downside:
dispatching can be implemented for concrete variants types (u8, char and so on),
type parameters, reference to a concrete type (&u8, &char), with a
correct lifetime, but **not** reference to a type parameter.

I don't know if this later case can (or will) be fixed.

This commit also implements tests for n-variants enums of concrete type.
Implementation of `Container`, type-argument varianted enums and
reference-varianted enums still have to be tested.