Better proc_macros, more documentation
Previous commit introced a derive macro which allowed to automatically
implement the `Dispatch` trait. This procedural macro was just a wrapper
to the declarative macro `implement_dispatch`. This commit introduces
manual implementation of the trait, instead of relying on previously
used declarative macro.

Additionaly, this macro is able to handle unnamed tuple variants.

This commit also adds some documentation for functions defined in the
`edisp-derive` crate.

Finally, a third crate was created, entitled `edisp-core`. Code is
organized as follows:
  - `edisp-core` contains declaration of `Dispatch` trait, implementation for
    for `std` enums and the `implement_dispatch` macro.
  - `edisp-derive` contains the procerural macro which allows to derive
    the `Dispatch` trait for user-defined enums.
  - `edisp` contains tests for the `Derive` procedural macro (these
    tests can not be written in `edisp-derive`), reexports important
    traits macros, and contains a lot of documentation.